+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorOneOf2Test.kt
@@ -87,7 +87,9 @@ open class TypeA(
     override fun hashCode(): Int =
             aaa.hashCode()
 
-    fun copy(
+    override fun toString() = "TypeA(aaa=${'$'}aaa)"
+
+    open fun copy(
         aaa: Long? = this.aaa
     ) = TypeA(aaa)
 
@@ -105,12 +107,12 @@ open class TypeA(
         override fun hashCode(): Int = super.hashCode() xor
                 xxx.hashCode()
 
+        override fun toString() = "A(aaa=${'$'}aaa, xxx=${'$'}xxx)"
+
         fun copy(
             aaa: Long? = this.aaa,
             xxx: String = this.xxx
         ) = A(aaa, xxx)
-
-        operator fun component1() = aaa
 
         operator fun component2() = xxx
 
@@ -128,12 +130,12 @@ open class TypeA(
         override fun hashCode(): Int = super.hashCode() xor
                 yyy.hashCode()
 
+        override fun toString() = "B(aaa=${'$'}aaa, yyy=${'$'}yyy)"
+
         fun copy(
             aaa: Long? = this.aaa,
             yyy: String = this.yyy
         ) = B(aaa, yyy)
-
-        operator fun component1() = aaa
 
         operator fun component2() = yyy
 
@@ -151,12 +153,12 @@ open class TypeA(
         override fun hashCode(): Int = super.hashCode() xor
                 zzz.hashCode()
 
+        override fun toString() = "C(aaa=${'$'}aaa, zzz=${'$'}zzz)"
+
         fun copy(
             aaa: Long? = this.aaa,
             zzz: String? = this.zzz
         ) = C(aaa, zzz)
-
-        operator fun component1() = aaa
 
         operator fun component2() = zzz
 
@@ -179,12 +181,12 @@ open class TypeA(
         override fun hashCode(): Int = super.hashCode() xor
                 qqq.hashCode()
 
+        override fun toString() = "D(aaa=${'$'}aaa, qqq=${'$'}qqq)"
+
         fun copy(
             aaa: Long? = this.aaa,
             qqq: String? = this.qqq
         ) = D(aaa, qqq)
-
-        operator fun component1() = aaa
 
         operator fun component2() = qqq